requirements.txt
pylint
selenium
instapy
psycopg2-binary
